long

+++ src/main/java/uz/maniac4j/IdGenerator.java
@@ -51,18 +51,18 @@ public class IdGenerator {
         return Response.ok(incrementAndGet()).build();
     }
 
-    public synchronized int incrementAndGet() {
-        int currentValue = 0;
+    public synchronized long incrementAndGet() {
+        long currentValue = 0;
         try (BufferedReader reader = new BufferedReader(new FileReader(file))) {
             String line = reader.readLine();
             if (line != null && !line.isEmpty()) {
-                currentValue = Integer.parseInt(line.trim());
+                currentValue = Long.parseLong(line.trim());
             }
         } catch (IOException | NumberFormatException e) {
             throw new RuntimeException("Failed to read number from file", e);
         }
 
-        int newValue = currentValue + 1;
+        long newValue = currentValue + 1;
 
         try (BufferedWriter writer = new BufferedWriter(new FileWriter(file))) {
             writer.write(Long.toString(newValue));
